AG-10004 - Working thumbnail generation.
